Added functionality for automated cleanup
