+++ src/models/AutomatedCleanupConfigs.ts
@@ -7,9 +7,12 @@ export interface IAutomatedCleanupConfigs {
 export class AutomatedCleanupConfigsCleaner {
     static cleanup(instance: IAutomatedCleanupConfigs) {
         return {
-            mostRecentLimit: Number(instance.mostRecentLimit),
-            cronSchedule: `${instance.cronSchedule || ''}`,
-            timezone: `${instance.timezone || ''}`,
+            mostRecentLimit:
+                Number(instance.mostRecentLimit) > 0
+                    ? Number(instance.mostRecentLimit)
+                    : 1,
+            cronSchedule: `${instance.cronSchedule || ''}`.trim(),
+            timezone: `${instance.timezone || ''}`.trim() || 'UTC',
         }
     }
 }
